cohnreznick/amortizationPro/requirements.txt
django-pyodbc-azure
djangorestframework
psycopg2-binary
django
Werkzeug
itsdangerous